+++ inventree_datanorm_plugin/tests/datpreis_test.001
